requirements.txt
pytest==8.3.3
pytest-playwright==0.5.2
pytest-html==4.1.1
PyYAML==6.0.2
requests==2.32.5
allure-pytest==2.13.5
allure-python-commons==2.13.5
Faker==40.1.0
